basic testing framework and move to a more modular composition
