+++ Gemfile
@@ -6,3 +6,6 @@ gem 'octokit', '4.1.0'
 gem 'redis', '3.2.1'
 gem 'resque', '1.25.2'
 gem 'sinatra', '1.4.6'
+
+gem 'rspec', '~>3.3.0',      :groups => [:development, :test]
+gem 'rack-test'
