+++ Rakefile
@@ -9,9 +9,6 @@ end
 
 desc "Run those specs"
 task :spec do
-  require 'spec/rake/spectask'
-
-  Spec::Rake::SpecTask.new do |t|
-    t.spec_files = FileList['spec/**/*_spec.rb']
-  end
+  require 'rspec/core/rake_task'
+  RSpec::Core::RakeTask.new(:spec)
 end
